[expo-ui] Move only single-platform components to namespaces.
