+++ packages/expo-ui/src/types.ts
@@ -1,14 +1,14 @@
-export type * from '../BottomSheet';
+export type * from '../swift-ui/BottomSheet';
 export type * from '../Button';
-export type * from '../ColorPicker';
+export type * from '../swift-ui/ColorPicker';
 export type * from '../ContextMenu';
 export type * from '../DatePicker';
-export type * from '../Gauge';
-export type * from '../Label';
-export type * from '../List';
+export type * from '../swift-ui/Gauge';
+export type * from '../swift-ui/Label';
+export type * from '../swift-ui/List';
 export type * from '../Picker';
 export type * from '../Progress';
-export type * from '../Section';
+export type * from '../swift-ui/Section';
 export type * from '../Slider';
 export type * from '../Switch';
 export type * from '../TextInput';
