+++ packages/expo-ui/swift-ui/BottomSheet/index.tsx
@@ -1,7 +1,24 @@
 import { requireNativeView } from 'expo';
-import { Dimensions, NativeSyntheticEvent, View } from 'react-native';
+import { Dimensions, NativeSyntheticEvent, StyleProp, View, ViewStyle } from 'react-native';
 
-import { BottomSheetProps } from '.';
+export type BottomSheetProps = {
+  /**
+   * Optional styles to apply to the `BottomSheet` component.
+   */
+  style?: StyleProp<ViewStyle>;
+  /**
+   * The children of the `BottomSheet` component.
+   */
+  children: any;
+  /**
+   * Whether the `BottomSheet` is opened.
+   */
+  isOpened: boolean;
+  /**
+   * Callback function that is called when the `BottomSheet` is opened.
+   */
+  onIsOpenedChange: (isOpened: boolean) => void;
+};
 
 type NativeBottomSheetProps = Omit<BottomSheetProps, 'onIsOpenedChange'> & {
   onIsOpenedChange: (event: NativeSyntheticEvent<{ isOpened: boolean }>) => void;
