+++ packages/expo-ui/swift-ui/Gauge/index.tsx
