+++ packages/expo-ui/swift-ui/Section/index.tsx
@@ -1,3 +1,4 @@
+import { requireNativeView } from 'expo';
 import { StyleProp, ViewStyle } from 'react-native';
 
 export type SectionProps = {
@@ -9,11 +10,16 @@ export type SectionProps = {
   style?: StyleProp<ViewStyle>;
 };
 
+const SectionNativeView: React.ComponentType<SectionProps> = requireNativeView(
+  'ExpoUI',
+  'SectionView'
+);
+
 /**
  * Section component uses the native [Section](https://developer.apple.com/documentation/swiftui/section) component.
  * It has no intrinsic dimensions, so it needs explicit height or flex set to display content (like ScrollView).
  * @platform ios
  */
-export function Section({ children }: SectionProps) {
-  return children;
+export function Section(props: SectionProps) {
+  return <SectionNativeView {...props} />;
 }
